Move ConfigurableTypeFactory to core package
