+++ money-core/src/main/scala/com/comcast/money/core/ConfigurableTypeFactory.scala
@@ -1,23 +1,7 @@
-/*
- * Copyright 2012 Comcast Cable Communications Management, LLC
- *
- * Licensed under the Apache License, Version 2.0 (the "License");
- * you may not use this file except in compliance with the License.
- * You may obtain a copy of the License at
- *
- *     http://www.apache.org/licenses/LICENSE-2.0
- *
- * Unless required by applicable law or agreed to in writing, software
- * distributed under the License is distributed on an "AS IS" BASIS,
- * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
- * See the License for the specific language governing permissions and
- * limitations under the License.
- */
-
-package com.comcast.money.core.internal
+package com.comcast.money.core
 
 import com.typesafe.config.Config
-import org.slf4j.{ Logger, LoggerFactory }
+import org.slf4j.{Logger, LoggerFactory}
 
 import java.lang.reflect.Modifier
 import scala.reflect.ClassTag
@@ -38,7 +22,7 @@ trait ConfigurableTypeFactory[T <: AnyRef] {
   def create(config: Config): Option[T] =
     findConfigValue(config, TYPE_KEY)
       .flatMap(findKnownType)
-      .map(_(config))
+      .map(_ (config))
       .orElse(findConfigValue(config, CLASS_KEY)
         .flatMap(createInstance(_, config)))
       .orElse(defaultValue)
