+++ money-core/src/main/scala/com/comcast/money/core/context/ContextStorageFilterFactory.scala
@@ -16,8 +16,7 @@
 
 package com.comcast.money.core.context
 
-import com.comcast.money.core.DisabledContextStorageFilter
-import com.comcast.money.core.internal.ConfigurableTypeFactory
+import com.comcast.money.core.{ConfigurableTypeFactory, DisabledContextStorageFilter}
 import com.typesafe.config.Config
 
 import scala.reflect.ClassTag
